+++ src/lib/route-data/index.ts
@@ -1,5 +1,0 @@
-export { LabelFunction, RouteData, ReadonlyRouteData } from './route-data-types';
-export { RouteDataInterface } from './route-data.interface';
-export { RouteDataRegistry } from './route-data-registry';
-export { RouteDataRegistryInitializer } from './route-data-registry-initializer';
-export { extractRouteDataFromRouterConfig, findMatchingRoutePath, getChildren } from './utils';