+++ src/lib/route/data/route-data-registry.ts
@@ -1,9 +1,10 @@
 import { RouteDataInterface } from './route-data.interface';
 import { Injectable } from '@angular/core';
 import { ReadonlyRouteData, RouteData } from './route-data-types';
-import { UrlHelper } from '../helpers';
-import { findMatchingRoutePath, normalizeRoutePath } from './utils';
+import { UrlHelper } from '../../helpers';
+import { findMatchingRoutePath, normalizeRoutePath } from '../utils';
 
+/** Manages extra data (such as a label) concerning routes of the app. */
 @Injectable()
 export class RouteDataRegistry<T extends RouteDataInterface> {
     private registry: Map<string, Readonly<T>> = new Map();