+++ src/lib/route/data/route-data-types.ts
@@ -1,6 +1,8 @@
 import { L10nTranslationService } from 'angular-l10n';
 import { Data, Params } from '@angular/router';
 import { RouteDataInterface } from './route-data.interface';
+import { InjectionToken } from '@angular/core';
+import { TerraKeyValueInterface } from '../../models';
 
 /** Type of method that returns a label based on the activated route's params, data and queryParams.
  * angular-l10n's L10nTranslationService can be used to provide multi-lingual labels */
@@ -13,6 +15,13 @@ export type LabelFunction = (
 
 export type RedirectedRoute = { redirected?: boolean };
 
+/** Specifies a set of mutable data related to certain routes. */
 export type RouteData<T extends RouteDataInterface> = { [path: string]: T };
 
+/** Specifies a set of immutable data related to certain routes. */
 export type ReadonlyRouteData<T extends RouteDataInterface> = { [path: string]: Readonly<T> };
+
+/** Injection token for the pre-extracted data of routes in the app. */
+export const ROUTE_DATA: InjectionToken<TerraKeyValueInterface<RouteData<RouteDataInterface>>> = new InjectionToken(
+    'route data'
+);