+++ src/lib/route/utils/route-children.ts
@@ -1,7 +1,7 @@
 import { Route, Routes } from '@angular/router';
 
 /**
- * This function returns the children of a given route wether they are lazy loaded or not
+ * This function returns the children of a given route regardless they are eagerly or lazy loaded.
  * @param route The route to which the children or the lazy loaded children should be returned
  */
 export function getChildren(route: Route): Routes | undefined {